+++ mayan/apps/ocr/icons.py
@@ -28,10 +28,6 @@ icon_document_version_ocr_download = Icon(
     driver_name='fontawesome-dual', primary_symbol='font',
     secondary_symbol='arrow-down'
 )
-icon_document_version_ocr_errors_list = Icon(
-    driver_name='fontawesome-dual', primary_symbol='font',
-    secondary_symbol='exclamation'
-)
 
 # Document type
 
@@ -41,9 +37,4 @@ icon_document_type_ocr_settings = Icon(
 icon_document_type_submit = Icon(
     driver_name='fontawesome-dual', primary_symbol='font',
     secondary_symbol='arrow-right'
-)
-
-icon_entry_list = Icon(
-    driver_name='fontawesome-dual', primary_symbol='font',
-    secondary_symbol='exclamation'
 )