+++ mayan/apps/ocr/links.py
@@ -3,13 +3,11 @@ from django.utils.translation import ugettext_lazy as _
 from mayan.apps.navigation.classes import Link
 
 from .icons import (
-    icon_entry_list, icon_document_type_ocr_settings,
-    icon_document_type_submit,
+    icon_document_type_ocr_settings, icon_document_type_submit,
     icon_document_version_ocr_content_delete_multiple,
     icon_document_version_ocr_content_delete_single,
     icon_document_version_ocr_content_view,
     icon_document_version_ocr_download,
-    icon_document_version_ocr_errors_list,
     icon_document_version_ocr_submit_multiple,
     icon_document_version_ocr_submit_single,
     icon_document_version_page_ocr_content_detail_view,
@@ -60,12 +58,6 @@ link_document_version_ocr_download = Link(
     permissions=(permission_document_version_ocr_content_view,),
     text=_('Download OCR text'), view='ocr:document_version_ocr_download'
 )
-link_document_version_ocr_errors_list = Link(
-    args='resolved_object.id',
-    icon=icon_document_version_ocr_errors_list,
-    permissions=(permission_document_version_ocr_content_view,),
-    text=_('OCR errors'), view='ocr:document_version_ocr_error_list'
-)
 link_document_version_ocr_submit_multiple = Link(
     icon=icon_document_version_ocr_submit_multiple,
     text=_('Submit for OCR'), view='ocr:document_version_ocr_submit_multiple'
@@ -90,12 +82,4 @@ link_document_version_page_ocr_content_edit_view = Link(
     icon=icon_document_version_page_ocr_content_edit_view,
     permissions=(permission_document_version_ocr_content_edit,),
     text=_('Edit OCR'), view='ocr:document_version_page_ocr_content_edit_view'
-)
-
-# Other
-
-link_entry_list = Link(
-    icon=icon_entry_list,
-    permissions=(permission_document_version_ocr,), text=_('OCR errors'),
-    view='ocr:entry_list'
 )